+++ pubspec.yaml
@@ -23,7 +23,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-
+  flutter_localizations:
+    sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -56,6 +57,7 @@ dependencies:
   flutter_launcher_icons:
   path:
   table_calendar:
+  intl_utils:
 
 
 dev_dependencies:
